+++ src/graphic/wordwrap.cc
@@ -336,7 +336,7 @@ void WordWrap::draw(RenderTarget& dst, Vector2i where, Align align, uint32_t car
 		if (where.y >= dst.height() || int32_t(where.y + fontheight) <= 0)
 			continue;
 
-		Vector2f point(where.x, where.y);
+		Vector2i point(where.x, where.y);
 
 		if (alignment == UI::Align::kRight) {
 			point.x += wrapwidth_ - kLineMargin;
@@ -353,9 +353,9 @@ void WordWrap::draw(RenderTarget& dst, Vector2i where, Align align, uint32_t car
 			int caret_x = text_width(line_to_caret, fontsize_);
 
 			const Image* caret_image = g_gr->images().get("images/ui_basic/caret.png");
-			Vector2f caretpt;
+			Vector2i caretpt;
 			caretpt.x = point.x + caret_x - caret_image->width() + kLineMargin;
-			caretpt.y = point.y + (fontheight - caret_image->height()) / 2.f;
+			caretpt.y = point.y + (fontheight - caret_image->height()) / 2;
 			dst.blit(caretpt, caret_image);
 		}
 	}
